refactor: extract agenda and task lists
